+++ src/main/java/com/test_task/repository/PositionRepository.java
@@ -3,7 +3,7 @@ package com.test_task.repository;
 import com.test_task.model.Position;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PositionRepository extends JpaRepository<Position, Integer>
+public interface PositionRepository extends JpaRepository<Position, Long>
 {
 
 }
