+++ src/main/java/com/test_task/repository/ProductRepository.java
@@ -3,7 +3,7 @@ package com.test_task.repository;
 import com.test_task.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Integer>
+public interface ProductRepository extends JpaRepository<Product, Long>
 {
 
 }
