+++ src/main/java/com/test_task/service/EmployeeService.java
@@ -27,7 +27,7 @@ public class EmployeeService
         return employeeRepository.save(employee);
     }
 
-    public Optional<Employee> findById(Integer id)
+    public Optional<Employee> findById(Long id)
     {
         return employeeRepository.findById(id);
     }
@@ -37,7 +37,7 @@ public class EmployeeService
         return employeeRepository.findAll();
     }
 
-    public Employee update(Integer id, Employee updatedEmployee)
+    public Employee update(Long id, Employee updatedEmployee)
     {
         return employeeRepository.findById(id)
                 .map(employee ->
@@ -54,7 +54,7 @@ public class EmployeeService
                 .orElseThrow(() -> new RuntimeException("Employee not found"));
     }
 
-    public void delete(Integer id)
+    public void delete(Long id)
     {
         employeeRepository.deleteById(id);
     }
