+++ src/main/java/com/test_task/service/PositionService.java
@@ -19,7 +19,7 @@ public class PositionService
         return positionRepository.save(position);
     }
 
-    public Optional<Position> findById(Integer id)
+    public Optional<Position> findById(Long id)
     {
         return positionRepository.findById(id);
     }
@@ -29,7 +29,7 @@ public class PositionService
         return positionRepository.findAll();
     }
 
-    public Position update(Integer id, Position updatedPosition)
+    public Position update(Long id, Position updatedPosition)
     {
         return positionRepository.findById(id)
                 .map(position ->
@@ -40,7 +40,7 @@ public class PositionService
                 .orElseThrow(() -> new RuntimeException("Position not found"));
     }
 
-    public void delete(Integer id)
+    public void delete(Long id)
     {
         positionRepository.deleteById(id);
     }
